pyproject.toml
[project]
name = "gamevault"
version = "0.1.0"
description = ""
authors = [
    {name = "parth",email = "parthsethia33@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "requests (>=2.32.4,<3.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "fastapi (>=0.115.13,<0.116.0)",
    "uvicorn (>=0.34.3,<0.35.0)",
    "jinja2 (>=3.1.6,<4.0.0)",
    "sqlalchemy (>=2.0.41,<3.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "databases (>=0.9.0,<0.10.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "passlib[bcrypt] (>=1.7.4,<2.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "itsdangerous (>=2.2.0,<3.0.0)",
    "selenium (>=4.33.0,<5.0.0)",
    "undetected-chromedriver (>=3.5.5,<4.0.0)"
]

[tool.poetry]
packages = [{include = "gamevault", from = "src"}]

[tool.poetry.scripts]
gamevault = "gamevault.cli:main"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
